Factor article image processing logic out of RoverArticleImageProcessingActor
